(temporarily?) removed a vim server guess
The idea was to run through the list of vim servers and ask each one
what file it was editing - if it ended in .tex, it was probably the
server we were looking for.  However, this proved problematic when
launched from within vim - the script would freeze when trying to run
the remote-expr.